Pull NodesFrame to own file
